[FE] refactor : profile api 영역 네이밍 수정, 디렉토리 정리
