+++ frontend/src/service/@shared/hooks/queries/user/index.ts
@@ -1,0 +1,1 @@
+export * from './useGet';
