+++ frontend/src/service/@shared/hooks/queries/user/useGet.ts
@@ -1,10 +1,24 @@
 import { useQuery, UseQueryOptions } from 'react-query';
 
 import { ErrorResponse } from 'service/@shared/types/index';
-import { GetMyReviewsResponse } from 'service/@shared/types/profile';
+import { GetMyReviewFormsResponse, GetMyReviewsResponse } from 'service/@shared/types/profile';
 
 import userAPI from 'service/@shared/api/user';
 import { QUERY_KEY } from 'service/@shared/constants';
+
+function useGetMyReviewForms(
+  queryOptions?: UseQueryOptions<GetMyReviewFormsResponse, ErrorResponse>,
+) {
+  return useQuery<GetMyReviewFormsResponse, ErrorResponse>(
+    [QUERY_KEY.DATA.REVIEW_FORM, QUERY_KEY.API.GET_MY_REVIEW_FORMS],
+    () => userAPI.getMyReviewForms(),
+    {
+      suspense: true,
+      useErrorBoundary: false,
+      ...queryOptions,
+    },
+  );
+}
 
 function useGetMyReviews(queryOptions?: UseQueryOptions<GetMyReviewsResponse, ErrorResponse>) {
   return useQuery<GetMyReviewsResponse, ErrorResponse>(
@@ -18,4 +32,4 @@ function useGetMyReviews(queryOptions?: UseQueryOptions<GetMyReviewsResponse, Er
   );
 }
 
-export default useGetMyReviews;
+export { useGetMyReviewForms, useGetMyReviews };
